requirements.txt
aiogram~=3.4.1
asyncpg~=0.29.0
pytest~=8.0.2
python-dotenv~=1.0.1
requests~=2.32.2